search: introduce RepoIndexOptions
Currently we just use "GetVersion" which allows us to abstract away how
we resolve a branch name to a commit SHA. We want to also introduce a
way to get a repository ID from the name. Instead of adding another
function, we introduce an option struct to carry these values.

The next commit will introduce repository ID.