search: sync access `calledSearchFilesInRepos` mock (#14674)
We call searchFilesInRepos twice in a normal search request. Once for files
and once for file contents. This means the mocks we use in tests need to
protect against concurrent access. CI picked up a test having race
condition. This PR switches us to using an atomic value.

Commit was created as a series of comby refactors:

  'calledSearchFilesInRepos := :[[a]]' 'calledSearchFilesInRepos := atomic.NewBool(:[a])'
  'calledSearchFilesInRepos = :[[a]]' 'calledSearchFilesInRepos.Store(:[a])'
  'if :[a~!?]calledSearchFilesInRepos' 'if :[a]calledSearchFilesInRepos.Load()'

Note: Technically we should validate which searchFilesInRepos we are calling
(ie for file paths or not). However, these sort of conditions are somewhat low
value. So rather than doing a bigger change I just fixed the race condition.