Reorganize reconciler action handlers
For the "undraft" action, I needed to create a lot of additional actionTypes, which led to a big bloat in the code. "reopenUpdraft", "updateUndraft", "undraft" all basically did the same, but needed to be separate actions.
The new structure builds up a list of actions that need to be taken. Then, it orders them in an execution order that makes sense (ie. not sync before update, update before undraft, so the latest PR description, .. will go out in the notification emails).
After that, all the actions are performed serially, but they don't all contain the upsert logic.
After all actions are done, the changeset is either upserted, or synced with the codehost, depending on what makes the most sense. Moving this logic out of the handlers makes the separate actions more composable, so I could build things like {reopen, undraft}.