Remove Campaigns beta flag
This removes the beta flag from campaign.

Yesterday, in a discussion with Chris, Quinn S., Christina and Rob we
decided that we shouldn't keep pushing that "remove beta-flag"-milestone
in front of us.

We think the product is ready to be used by more customers and our
primary goal is to grow adoption. That means it's no longer in beta.

Of course we'll still iterate and refine and improve campaigns, that's
out of the question.

So, let's get rid of this artificial barrier and milestone and focus on
the *actual goal*: getting campaigns into the hands of more customers.