Add support for draft changesets in reconciler (#14807)
* Implement publishDraft action in reconciler

In this commit, I've done the plumbing necessary to make draft changeset publishing possible.
It contains:
- New method to determine if a codehost has a capability, and a config map that makes a global place to document them.
- Introduce a DraftChangesetSource, so that only ChangesetSources that actually support Draft mode need to implement it.
- Adds reconciler tests for publishDraft.
- Adds more lightweight determineAction tests, that can assert the reconciler action has been determined correctly.

* Implement undraft reconciler action

This adds the second bit to the reconciler: Moving a changeset from being a draft to being fully published.

In this chunk:
- Extend the DraftChangesetSource by a UndraftChangeset method. This is implemented on the GitHub source and client.
- Extend the reconciler integration tests and the newly added determineAction test suite.
- Add a temporary undraft-update reconciler action. This will not stay, and will be gone in the next commit already.
    The main purpose of adding it here is to make it work, so I can validate it does what it's supposed to and can write tests,
    that should not break in the refactor in the next commit. This state is currently flawed and doesn't support all edge-cases where
    the changeset might need to be moved out of draft mode. Some of them are
        - attributes don't change (likely to happen)
        - reopen without update won't respect the update
        - does unnecessary updates to the changeset on the codehost, when only the published flag changed

* Reorganize reconciler action handlers

For the "undraft" action, I needed to create a lot of additional actionTypes, which led to a big bloat in the code. "reopenUpdraft", "updateUndraft", "undraft" all basically did the same, but needed to be separate actions.
The new structure builds up a list of actions that need to be taken. Then, it orders them in an execution order that makes sense (ie. not sync before update, update before undraft, so the latest PR description, .. will go out in the notification emails).
After that, all the actions are performed serially, but they don't all contain the upsert logic.
After all actions are done, the changeset is either upserted, or synced with the codehost, depending on what makes the most sense. Moving this logic out of the handlers makes the separate actions more composable, so I could build things like {reopen, undraft}.