search: report function for aggregator and one mutex (#15046)
Our intention is to have a uniform interface on how results are bubbled
back. This will the be morphed into an API used by streaming.

Additionally we use a single mutex for aggregator. This is easier to
reason about since there is only a single mutex. Additionally this is
likely more performant since we do less high frequency fine grained
locking.