Add error codes to campaign errors
Some of our resolvers return errors very purposefully, and we did document that _some_ error will be returned, but not which or how to identify this specific error.

I tried making more clear what the error actually looks like and that people using the API can assert on them by adding an error code to them, through an Extensions field that graphql-go provides us with.