search: aggregator merges file match results in report (#15048)
This is a refactoring that moves the merge logic from the invidual
functions to the new report function. Additionally we implement true
merge logic, rather than assumptions based on if we have a symbol vs
filematch result.

There is one potential downside. We now allocate an intermediate
"[]SearchResultResolver" slice. In practice I believe this won't be that
big of an issue since the size of the interface "SearchResultResolver"
is tiny compared to the size of a "FileMatchResolver".