gitserver: dynamically sample honeycomb based on command (#15269)
We are hitting our limits again. By far rev-parse is our most common
command. We are currently sending 3500 requests to honeycomb a second
just for rev-parse. The next most common is log, which we send 280
req/s.

This commit dramatically scales down the amount of data we send for
rev-parse. Most of this information is not that useful. Note: honeycomb
does store in the samples the sample rate. This allows us to effectively
get count graphs that we can compare.