search: add events to MonitorTrigger
This PR completes the query-part of the GraphQL schema for code monitors.
By requirement, we have to be able to link action events to the events
of their associated trigger. Hence each trigger event is a link of a single
MonitorEvent (the event of the trigger itself) and a list of events from
the actions triggered by the trigger.