git: introduce a scanner for parsing the output of git log
This is an internal implementation detail of how we run git
log. Currently we read all the data from the output of git log, and then
process it (and run further git show commands). Instead we will be
making it based on an io.Reader to make it amenable to streaming (and
more efficient!).

This commit just introduces the scanner, but hides it behind the current
parseCommitsFromOnelineLog function. The next commit will remove that
function and directly use the scanner.