git: directly run scanner over git log reader
This avoids the need to call "ioutil.ReadAll". Instead we can read the
output of log as it streams out. This makes us one step closer to making
a pure streaming diff/commit search.

This commit needs to handle some of the work done by
readUntilTimeout. For example marking timeout as incomplete, as well as
ignoring empty repositories.