git: test revision not found in diff search
This exposed that our "RevisionNotFoundError" logic is broken. I don't
think it ever worked, so I have removed it for now. We resolve revisions
in the graphqlbackend before calling git.RawLogDiffSearch so you can't
trigger this code path in practice. So the added test case is for
completeness, rather than testing a practical code path.

I will follow-up with a more general fix to gitserver.StdoutReader which
understands common git error lines.