git: test error cases in RawLogDiffSearch (#15334)
We increase the test coverage of diff/commit search. In particular the
testing added focuses on error cases that can occur. Currently there is
error code inspecting output of git in multiple parts of the log parsing
code. This made it tricky converting the function to be stream based. As
such these are added to ensure we do not regress.