git: use Scanner instead of Buffer for large refs
bufio.NewReader has a default size of 4kb. This means the maximum token
we could read has a size of 4kb. However, refs can be much larger than
that (although rarely are). So we switch the implementation to Scanner
which supports growing the buffer (upto 64kb).

This change additionally removes skipping over empty parts. This won't
happen. Even in a timeout, we will instantly return the timeout
error (rather than partial buffered data).

This implementation suggests we should change the interface to look like
Scanner (instead of a scan function). I am avoiding that for now since
it introduces extra state for not much benefit.