repo-updater: handle error from remoteRepoSync
Currently production is panicing if we try to fetch a repository which
is blocked. In particular for youtube-dl repo GitHub is returning:

  status 451: Repository access blocked

This leads to us returning an error and a nil repoResult. We then try to
access the nil repoResult. This adds in the appropriate early return to
avoid reading repoResult.