Make sure test fails if assertion is wrong
This is the result of this comment: https://github.com/sourcegraph/sourcegraph/pull/15068#discussion_r517210400

The test wasn't wrong, but this changes it to enforce the license check
in possibly every mutation.

That has the advantage that when we add the license check is added to
another mutation a test case here will fail and alert us to add the
`wantLicenseErr: true`, to make the desired behaviour explicit.