git: factor out logDiffCommonArgs (#15378)
We have these as inline functions in an already large
function (RawLogDiffSearch). They can instead be functions which take in
the options and generate the common arguments.

Note: its clearer now that we can compute the common args once. However,
the work it does is small and I plan on splitting this function up more.