git: scan git log output instead of readall then process (#15348)
This is an internal implementation detail of how we run git
log. Currently we read all the data from the output of git log, and then
process it (and run further git show commands). Instead we will be
making it based on an io.Reader to make it amenable to streaming (and
more efficient!).

This change additionally removes skipping over empty parts. This won't
happen. Even in a timeout, we will instantly return the timeout
error (rather than partial buffered data).