license_expiration: Check all subscriptions (#15483)
The previous code would return in the for loop where all subscriptions
were checked. A common case for that to happen was when a subscription
had less than one license.

Instead, we extract the code for checking each subscription into a
separate function, and we call that for each subscription, even in the
face of errors in the previous one.