conf: read conf source first and then wait for updates
This changes the conf server so it reads the configuration at startup
and _then_ waits for updates/jitter-timer.

Before this change `frontend` would hang between 0-5 seconds (max time of
jitter) when starting up, because it was waiting for the configuration to
be read in this line:

https://github.com/sourcegraph/sourcegraph/blob/6a14d3cd1595f6576e64a74c98834bdc80710886/cmd/frontend/internal/cli/serve_cmd.go#L180

Since at the point in time when the `watchSource` goroutine is kicked
off, we already have setup the conf package (see [1]), it's okay to read
the configuration and _then_ wait for the jitter, especially since the
jitter could be 0seconds anyway.

[1]: https://github.com/sourcegraph/sourcegraph/blob/6a14d3cd1595f6576e64a74c98834bdc80710886/cmd/frontend/internal/cli/serve_cmd.go#L128-L133