gitlab: Always set token in clone URL even if public
This change was triggered by a customer running into an error when
trying to use campaigns on a public repository. `gitserver` couldn't
push the commit because the credentials in the remote URL were missing.

Turns out that we don't save the token for public GitLab repositories,
_even if we have one available_.

The problem is that users can't use campaigns with public Gitlab
repositories until we add user tokens and they then add a separate token.

That problem doesn't exist with GitHub or Bitbucket Server where we
always set the token.