graphqlbackend: concurrently fetch settings from DB (#15442)
On Sourcegraph.com it takes 35ms to compute a users settings. The bulk
of that time for me is 7 round trips to the database fetching the
settings for each subject. The settings are then merged. This commit
fetches the settings for each subject concurrently.

We introduce goroutine.Bounded. This is a helper which provides
goroutine runner which is bounded by a semaphore. Additionally it will
capture errors returned for convenience.

Co-authored-by: Erik Seliger <erikseliger@me.com>