search: Progress API in webapp (#15430)
This is an implementation of converting the proposed Progress API into
the existing GraphQL search results. Note the backend does not send this
event yet, it will be done follow-up PR.

This is not the same as the proposal. The following changes have been
made after feedback:

Removing counts
^^^^^^^^^^^^^^^

Lots of feedback over if shards is a good word to communicate. The
progress API is instead updated to just communicate "copy" for
implementing the UI. The only places we will now expose counts is total
number of repositories and in the messages for skipped items.

Done
^^^^

The boolean field Done is added to indicate the progress update is the
final/last one. This is useful for affecting how we communicate
numbers. Until we have a final progress update all communication must
assume result sets can change.

Skipped
^^^^^^^

Skipped now has a title and a message. This matches the mocks in
Figma. Additionally suggested query now has a title and what should be
added to the query.

Co-authored-by: Juliana Peña <me@julip.co>
Co-authored-by: Loïc Guychard <loic@sourcegraph.com>