git: treat rawLogSearch as an io.ReadCloser
I spent the last 1h20m experimenting with changing the interface of how
we send results from rawLogSearch to rawDiffSearch. It become very
tricky to have a nice interface with channels which correctly handled
cleanup / error handling.

I ended up settling on a much cleaner interface which is based on the
idea of not hiding we are running a command which has an
io.ReadCloser. So instead we move the logic of setting up the log
command into its own function, but our main function is responsible for
creating the command since it knows best when to call close to
cleanup.

Once we have done that, we can also move to treating the output like a
Reader/Scanner/Iterator. This commit removed the batching to make the
implementation clearer, the next will add it back.

This interface is starting to feel good.