git: properly respect debounce time in batching
This was bothering me. If we had a search which matched very few
commits, we would end up blocking for potentially a very long time,
leading to bad behaviour. This feels like the common case, so I spent
some time ensuring the pipeline never stalls.

The main insight is that as soon as you create a goroutine, you need
some sort of cleanup/close function to signal cleanup of resources (the
goroutine you created). This function likely deservers a type now,
rather than relying on closures and inline function
definitions. However, this works :)