graphql: Record queries that trigger getCommit calls
This is an attempt to debug #15392 and find what GraphQL queries trigger
the `getCommit` codepath, which leads to a spike of git log calls in git
server, that triggers that alert.

In the future, we should invest in working tracing, rather than only
per-request activated tracing, which doesn't allow us to retroactively
debug situations like these.