git: stream output from log in batches into show  (#15386)
This greatly improves the perfomance of diff/commit search by streaming
the log output into the part which does git show. This also makes it
straight forward to change the API for RawLogDiffSearch into a stream
based API.

We want to pass multiple commits to git show to amortize the cost of
starting up git / RPC costs. So we batch up the commits returned from
git log. However, the time between git log outputting a commit and the
next is unbounded. We scan over the io.Reader returned from gitserver
for running the git log command. Because it's an io.Reader, we don't
have a way to timeout a read. As such we wrap our scanner in a goroutine
to send down the scans over a channel. This allows us to implement the
timeout behaviour of returning what has been fetched so far.