git: remove shorter deadline hack for git show
When we did one big batch of requests for commit search, we split up the
time as such:

1. git log gets 50%
2. git show gets 40% (80% of 50%)
3. git resolve-ref gets 10% (20% of 50%).

Now that we stream commits in batches from git log into git show, we
don't need to divide up the time. We still had "step 3" in the above in
our code. This commit removes that. Additionally git resolve ref only
ever does one request, which is finding out what HEAD points to. So
leaving 20% time for that was over the top before.