search: connect action resolvers to db
This PR replaces the stubs for the action resolvers.

The core changes related to the resolvers are all in `resolvers.go`.

However, most of the changes in this PR are related to new tests.
Especially with the new `TestQueryMonitor` we reach a coverage of
`71.9%` for code monitoring. Like before, I shamelessly copied a lot of
ideas and structure from campaigns, EG I factored types used in tests to
the `apitest` package, as well general purpose test helpers to
`main_test`.