graphqlbackend: searchCommitsInRepoStream has real streaming
We switch to using RawLogDiffSearchStream which streams out the
results. The main functional change here is we can send back more than
FileMatchLimit results. We instead just stop searching once we hit the
limit and send everything we did manage to find. This is similiar to how
it is done in other parts of our search aggregator.