graphqlbackend: factor out searchCommitsInRepos
The functions searchCommitLogInRepos and searchCommitDiffInRepos are
identical, except for which function they call out and some naming used
in tracing/errors. This commit factors out the common functionality and
passes into it the differences between the two.

This is motivated by reducing the amount of work we need to do for
streaming as well as just removing silly duplication.