search: add stub for UpdateCodeMonitor
This PR adds a stub for `UpdateCodeMonitor`. Once we align on the schema
I will remove the stub in a follow-up PR.

Note that the signature of `CreateCodeMonitor` is slightly different from the
signature of `UpdateCodeMonitor`. I think it might be a good idea to
align both, IE we should change the signature of `CreateCodeMonitor` and
nest the meta-fields (description, enabled, ...) of a code monitor on the
same level as trigger and actions. This way we could reuse the types in
the schema which makes the whole thing less brittle.