graphqlbackend: Lazy load git commit for each hunk
This change makes it so `hunkResolver` lazy loads the `git.Commit` of a
hunk from gitserver, instead of eagerly doing it for all hunks.

This functionality was already built in `GitCommitResolver`, we were
just not using it. Additionally, for use cases where only the commit URL
is requested, that resolver doesn't call out to gitserver, which fixes
the current problem with the git blame queries coming from https://github.com/sourcegraph/sourcegraph-git-extras

Fixes #15392