graphqlbackend: Add searchResolver.SetResultChannel
This is how our streaming and our batch interface co-exist. When this is
set, it exposes a way to stream out results as we collect them. We set
it in our from our streaming REST handler.

This is not our final design. For example this doesn't allow us to
stream out things like dynamic filters or take into account
AND/OR. However, streaming is behind a feature flag for now, so this is
to make it visible in the browser.