Faster compute of burndown charts (#15756)
Turns out for the burndown chart we really don't need all events in the PR timelines. Given that, we can add a filter by kind and sometimes drastically reduce the amount of events that need to be processed. In case of my 9k changesets campaign this reduces it to ~30% of the initial amount of events.

Before:
Required changesets: 9045, events: 106843, request duration: 5.34s

After:
Required changesets: 9045, events: 31752, request duration: 4.10s