repo-updater: Handle Ctrl-C correctly
I don't know when it started, but it seems like repo-updater ignores
Ctrl-C and goreman needs to kill it after a 10s wait.

This change here switches the code to use the `internal/httpserver`
package and the `goroutine.MonitorBackgroundRoutines` function that
listens to the signals and then shuts down the server.