e2e: allow users who aren't changing username to update other metadata
Normal users were previously blocked from changing metadata like `displayName` or `avatarURL` if `auth.enableUsernameChanges` was false[0]. This was because the UI always includes the `username` value in its payload, which triggered the check incorrectly on the backend and blocked an action that should have succeeded.

This updates the E2E tests to reflect what we actually want to test:

- if the site config prohibits username changes, attempting to change the element will fail and we'll see the same username after submission (even if other fields changed)
- if the site config permits username changes, the element will be enabled and any/all of the fields can be modified by the user on submission

The frontend suite covers cases like hand-crafted GraphQL requests that could potentially trigger the site configuration error.

ref https://github.com/sourcegraph/sourcegraph/issues/15877

[0] https://github.com/sourcegraph/sourcegraph/pull/15693