search: disable nginx buffering for streaming search (#15964)
This was a fun one to track down. Nginx buffers responses before passing
them on. There is a header you can set on a response to disable this
behaviour. Before this change streaming search would often only see the
first result once the full search had been complete. Now results are
shown as soon as they are found.