monitoring: Change syntect_server owner to Cloud team (#16083)
The code intel team never really owned this service. @slimsag built this
and mostly maintains it, but we have only team ownership of services, so
it makes sense to put this under Cloud / Backend Infra / Backend
Platform, because it's a shared service, like gitserver.

Time to learn some Rust I guess. :D