Do not try to parse raw ChangesetSpecs as YAML
In https://github.com/sourcegraph/sourcegraph/issues/16081 we discovered
that the YAML parser in `go-yaml` can choke on control characters, which
can be found in a diff.

Leaving aside whether that _is_ a bug in the YAML parser or not, it
doesn't make sense that we try to parse the raw changesetspec as YAML,
since it'll never be YAML.

So, with https://github.com/sourcegraph/campaignutils/pull/7 we now have
the ability to only parse and validate the input as JSON without going
through the YAML parser.

That in turn fixes #16081 on the server.