gitserver: get remote URL from database
Currently gitserver stores the remote URL in a repositories
.git/config. It is informed of this URL lazily. Some requests include
the "URL" field. When this field changes, we update the
configuration. This leads to hard to understand semantics in clients,
for example CachedGitRepo, GitRepo and CommandRetryer all exist for
different strategies of obtaining the remote URL.

We now have the CloneURL stored in the DB. So instead of gitserver
sometimes being told the URL, gitserver can directly ask the DB for the
URL. This commit implements that by introducing a connection to the
database in gitserver and updating all call sites which read the URL to
instead ask the database.

Co-authored-by: Tomás Senart <tomas@sourcegraph.com>