Gitlab: Make SetWIP and UnsetWIP work for both Draft and WIP prefixes (#16101)
Depending on the GitLab version, both of these can be used to mark something as "work in progress". To be more resilient about this. For example when doing "undraft" and someone changed the title on the codehost manually to be draft: .., we wouldn't actually have undrafted, so we should allow both in the logic.