Always refetch gitlab merge request data after mutation (#16100)
Whenever we modify the merge request on the codehost, there will potentially be new pipelines, new notes or changes to the MR itself. Before this change, sometimes after a reconciler action the UI state was inconsistent at first, like CI status not reported correctly for example.
This synchronizes the behavior with what we do for GitHub.
Nuance: GitHub allows us through graphql to get all the fields we need from a single GraphQL query whereas for GitLab it'll be multiple. I believe that is fine though, because
- we don't have another way around it
- for GitHub we consume 3-4 rate limit tokens with one request, so GitLab with 3 requests as well won't be standing out really