Restore backward compatibility with GitHub enterprise 2.20 (#16042)
This PR restores compatibility with GitHub enterprise 2.20, which seems to be still used by customers. I think once we deprecate support for that version we should revert this PR as it's kind of a mess.
We do so by fetching the instance version first and then constructing the GraphQL queries based on version constraints.