search frontend: render streaming search results (#16097)
Renders search results for streaming search!
This still uses the GraphQL data types. This will be later changed as @keegancsmith is going to update the API so conversion to GraphQL types is no longer required (primarily adding a string type to each result)

Note that lots of features are still missing here (eg. loading state, footer, etc). This is meant to be the initial work to render results.

There is a bug in storybook tests where not all code excerpts are being rendered: #16095 This doesn't happen in the real webapp.