Don't try to update merged changesets on reapply (#16128)
Previously, this would have tried to modify merged changesets (push, sleep, sync). I think we should not allow that, since it will be a noop or fail and costs rate limit tokens.
This can happen when I merge a change against a repo and then the campaign yields a NEW diff for that repo. (Ie, just append to a file). After it's merged, my command would append, hence the diff changes and the reconcilerOperations are {push, sleep, sync}. That actually updates the branch on the remote even, but then silently doesn't do anything as close is a noop on a merged PR.

This is the first find from the more transparent reconciler operations we'll get with the delta API. :)