Remove campaigns => changeset mapping
This is a duplicative source of truth to the IDs stored in the campaign and can cause inconsistencies that seem hard to debug. Also, selfishly it gives me the benefit that I can just upsert all changesets in the ApplyCampaign method, also to-be-closed/detached ones, because there is no need to distinguish them from attached ones anymore.