search: streaming final progress event
This is our initial support for progress events. Currently it is
partially implementing all the reasons we could skip. More invasive
changes are required to our graphqlbackend to report more. Additionally
it is only sending the progress event once at the end of the
stream. This will also require more changes to graphqlbackend.

Follow up commits will add progress events during search as well as a
implement all skipped reasons.