wip re-using observe message to send progress
This is a non functional commit which shows where my knowledge ends for
observables. I want to make it such that done now contains a progress
payload. I want to emit that then close up the observable. I tried
factoring out the observable bits (foo) then got stuck with how to use
it in the message handler for done.

Help plz