search: generate progress events in graphqlbackend (#16349)
As we improve on progress, we will need more access to "search
internals" which currently live in the graphqlbackend. I wanted to avoid
this, but for practicality purposes this is best.

To accomplish this we factor out the search progress types into a shared
package. The types have been exported as well as documented.

We have also added some golden tests.