vcs/git: Remove commandRetryer
This changeset removes the jurassic `commandRetryer` struct and logic,
that was used by `git.ResolveRevision` and `git.GetCommit`. With
gitserver now always knowing the clone URL for a given repo name it
receives, there's no need for the client side to have the logic to
acquire the remote URL if needed.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>