extensions: convert Range classes to plain Range when merging hovers (#16366)
Now that we use Range in codeintellify, we need to make sure that we're passing it the expected type in the web app.

Range lost its start and end getters as it was cloned to the main thread. This PR adds a check for the prototype of the range to be merged. If it's an instance of the Range class, convert it to a plain object.